[JP-06][API] Updated pagination types
