+++ src/ts/api/filters.api.ts
@@ -1,6 +1,6 @@
 import { yourEnergyApi } from './base';
 import { Filter, GetFilterParams } from '../types/filters';
-import PaginatedResponse from '../types/response.api';
+import { PaginatedResponse } from '../types/pagination';
 
 const getFilters = async (
   params: GetFilterParams
