+++ src/ts/types/pagination.ts
@@ -1,0 +1,13 @@
+type PaginationParams = {
+  page?: number;
+  limit?: number;
+};
+
+type PaginatedResponse<T> = {
+  page: number;
+  perPage: number;
+  totalPages: number;
+  results: T[];
+};
+
+export { PaginationParams, PaginatedResponse };
