--- src/ts/types/params.api.ts
@@ -1,6 +1,0 @@
-type PaginationParams = {
-  page?: number;
-  limit?: number;
-};
-
-export default PaginationParams;
