--- src/ts/types/response.api.ts
@@ -1,8 +1,0 @@
-type PaginatedResponse<T> = {
-  page: number;
-  perPage: number;
-  totalPages: number;
-  results: T[];
-};
-
-export default PaginatedResponse;
